+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/SpecVersion.java
@@ -20,6 +20,7 @@ import tech.pegasys.teku.spec.config.SpecConfigMerge;
 import tech.pegasys.teku.spec.logic.DelegatingSpecLogic;
 import tech.pegasys.teku.spec.logic.SpecLogic;
 import tech.pegasys.teku.spec.logic.versions.altair.SpecLogicAltair;
+import tech.pegasys.teku.spec.logic.versions.merge.SpecLogicMerge;
 import tech.pegasys.teku.spec.logic.versions.phase0.SpecLogicPhase0;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitions;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsAltair;
@@ -69,7 +70,7 @@ public class SpecVersion extends DelegatingSpecLogic {
 
   static SpecVersion createMerge(final SpecConfigMerge specConfig) {
     final SchemaDefinitionsMerge schemaDefinitions = new SchemaDefinitionsMerge(specConfig);
-    final SpecLogic specLogic = SpecLogicAltair.create(specConfig, schemaDefinitions);
+    final SpecLogic specLogic = SpecLogicMerge.create(specConfig, schemaDefinitions);
     return new SpecVersion(SpecMilestone.MERGE, specConfig, schemaDefinitions, specLogic);
   }